Introduce start event internal model from bpmn xml
Initialization of the BPMN Shape model
Initialization of the XML & Json parsers for BPMN 2.0

Implementation relies on
- fast-xml-parser -> xml data to json
- json2typescript -> json to TS objects

Manage
- empty process
- process declared as array or as single object
- multi start events

Build infrastructure changes

Introduce mocha/chai tests and run them with the GitHub Build Workflow
Introduce nyc test coverage

tsconfig: sourceMap to true
sourceMap option is needed to be true to make Mocha to correctly picking up line
numbers and also it helps Nyc to know the exact line numbers for test coverage
as well

mocha npm run: setup ts node compiler options to use module:commonjs
source: https://stackoverflow.com/a/44343768
see also https://journal.artfuldev.com/unit-testing-node-applications-with-typescript-using-mocha-and-chai-384ef05f32b2
The important parts are the module and moduleResolution. The value for module is
usually left as commonjs, but for code splitting to work, we need to change it
to esnext. And this will cause problems for our unit tests. To fix this, we need
to change the module back to commonjs when running mocha with ts-node

About the 'test' script in package.json
The syntax with triple backslash is required to make it run on Windows
Solution taken from https://github.com/kentcdodds/cross-env/issues/30#issuecomment-466328651