Refactor: detection based on ShapeBpmnElementKind enum only (#103)
This change will simplify work to do to support new BPMN element types.
Previously, we had to modify the `ShapeModelConverter` in addition to the `ShapeBpmnElementKind` enum. The enum change is mandatory because this is how we identify the element for rendering.
But having to change the converter was a pain and led to code duplication.

The converter now generically considers all elements of the enum to apply the conversion. The BPMN containers (lane for instance) are still managed manually as they require a dedicated processing.

Also introduce a new doc which explains briefly how to support a new BPMN element type.