[REFACTOR] ProcessConverter: simplify event definitions retrieval
Previously, for a given 'bpmn element', we retrieved a list with all event
definitions that exist in the spec with the number of occurrences associated to
the 'bpmn element'.
It was not obvious that this list contained all definitions with some occurrence
counters set to zero:
  - generally, the intuition of the reader is to have only definitions for non
  zero counts.
  - in addition, the method documentation confirmed this intuition 'Get the list
  of eventDefinition hold by the Event bpmElement'. It never talked about zero
  counters
Second, this required an extra filtering to only get the definitions actually
hold by the bpmn element (only keep definitions with non zero counter).

The new implementation now returns only a list of event definitions with non
zero counter i.e. the event definitions declared in the BPMN definition with
their occurrence count. This method is in charge of the zero counter filtering,
letting the client code cleaner.