[DOC] move all development info to the contributing guide
We decided to put most of development information in this guide, not in the
README.
The release part of the contributing guide was wrong whereas it was correct in
the README which was confusing.
Also add contributing guide and doc site development section cross links for a
better navigation and understanding.