fix: compensation event definition name
The compensation events were already displayed because its name was wrong in the
ShapeBpmnEventKind class. As a consequence, it was detected as a None event.
In the xml, it is identified by: <bpmn:compensateEventDefinition ...>
Notice that the issue previously also occurred with the End event, but hasn't
been detected earlier.