[REFACTOR] generalize icons management
Introduce a declarative way for the bpmn event render icon method selection
This uses a more functional way of doing things instead of having several
if/else or switch which brings more clarity and will simplify extensions.

Simplify how we handle catch/throw icon render
This is now hold by the Shape class itself allowing icon render method to easily
know if a Throw or Catch icon has to be painted.
Prior this change, we had to pass argument to several methods which made the
code hard to follow.

Introduce a single place to manage events currently displayed with a arbitrary
fill color prior we implement their final icon.
This was prior done at several places which make the code hard to follow and was
error prone: we have broken the render of such BPMN events several times in the
past.