fix computeStyle: make it consider isInterrupting when false
When false, the isInterrupting was undefined in the computed style and in that
case, the shape rendered interrupting boundary event instead of non interrupting
(expected behaviour when undefined).

isInterrupting was correctly set in the style values map, but when computing the
final style string, isInterrupting with false boolean values were omitted. This
is because of the values filter function used which filters undefined string, 0
number and false boolean.
Using string instead of boolean as value fixes the issue.