[TEST] check label values in mxgraph model e2e tests (#348)
This ensures that the BPMN element name are correctly set in the mxGraph model
prior being rendered.

In addition, simplify mxGraph model check functions signature to make the introduction
of text label checks easier. The functions previously have several arguments and
every time we needed to check new things, we added new arguments.
We now use interface arguments that will be enriched when needed instead of
changing the function signature.