[REFACTOR] canvas support scale on Y direction
The new supported scaling direction is currently not used by client code,
so this commit doesn't change any rendering behaviour. This will be managed
later.
This change is part of a global refactoring that aims to rationalize render
methods across the custom Canvas and the IconPainter.

Custom canvas used for icon/markers
  - rename `MxScaleFactorCanvas` into `BpmnCanvas` as it is intended to not only
  manage scaling in the future
  - introduce a single parameter constructor to support more and more
  configuration features (avoid signature change every time and large amount of
  parameters)
  - manage translate with scale factor
  - move to a shape subfolder as it is used for shape rendering

Configuration interfaces: start moving them to a dedicated file
  - extract the rename `IconStyleConfiguration` from `IconPainter`
  - introduce `IconStyleConfiguration` for all icon configuration (including
  height, width, ratio from shape, and more in the future)
  - extract `Size`