update non regression resource: use fitOnLoad to render diagrams
The dimension of the diagrams are now larger, so unexpected changes will have
more impact on the rendering (more differences)
Compared to the noise introduced by the running environment that is different
from the one used to create reference image snapshot, the proportion of the
differences will be higher than before.
So, this should help the diff tool to highlight them more easily (higher diff
percentage in that case
Notice that the noise is mainly introduced when there are labels on the diagram
(font rendering).