[TEST] improve visualization non regression tests
Add and update tests
  - labels: specific position for some label tasks, use line breaks
  - message flows
  - pools and lanes, including vertical orientations

Introduce a way to configure the diff snapshot threshold by tests and os (only
needed for BPMN diagrams with labels). This allows to configure the lowest
threshold to make the tests pass on all environments (including GitHub
Workflows).

Change the bpmn content passed to the page
We currently face issue when the BPMN content is too large. To make this issue
happen less often, we now linearize the xml to reduce the text length.
We also url encode the bpmn content to ensure xml entities are retrieved
correctly by the code in the page.

Demo update for better tracking
Log the bpmn received from url and log prior and after BPMN loading.