[TEST] Introduce BPMN rendering non regression visual tests (#537)
Introduce tests that will ensure non regression in BPMN element rendering.
Several BPMN diagrams are loaded with a page, a screenshot of the rendering is
done and compared to a reference image. In case of difference, the test fails
and the resulting diff image is generated for review.

Technical details
  - Use 'jest-image-snapshot' with the 'ssim' support for image diff.
  - Use dedicated html page for non regression tests: no button, the diagram fit
  the whole viewport after it is loaded.

Note: the implementation is mainly based on an inital poc previously done, see
GH #523.