[TEST] improve 'events' visual tests
Use a BPMN diagram that contains the whole events of the BPMN palette. So, we
now test all elements that the lib is currently supported, and we are prepared
to test remaining events to support in the future.

As the BPMN diagram is too large for the load method using the url query param
in the 'visual testing' page, add a new loading method:
  - allow the test page to load a BPMN diagram using a query param providing the
  url to the file to load (via the Fetch API)
  - copy the file from the test fixtures folder to a folder that is served by
  the test server
  - pass the relative path to the newly copied file as url parameter to make the
  page load this file
  - update the 'visual testing' page to display the BPMN file fetching errors

In addition, refactor and share functions used to load the BPMN content in e2e
tests.