[TEST] avoid false positive rendering test errors (#565)
When running with the macOS GitHub Workflow, the 'events' test is often marked
as failing. The screenshot is related to a blank page.
This occurs because the machine is slow and the test takes the screenshot prior
the rendering completes.
To avoid this, we now require some svg nodes to be available in the DOM: they
are created by mxGraph during the BPMN rendering.