[DOC] improve Contrib guide (testing and bpmn how to)
Clarify the documentation about tests and bpmn support how-to
- the Asciidoctor/html doc only contains architecture details
- general development information are in CONTRIBUTING.md
- development details are in dedicated markdown doc within the repository

This should avoid going back and forth between the html documentation and the
contributing guide, allowing contributors to have all information at the same
place within the repository, close to the source code.

The tests are now explained in a dedicated page, the BPMN support how-to
documentation refers to it instead of just describing the tasks to do without
context.