[REFACTOR] Avoid global states in json parsers (#721)
We previously had global instances shared by json parsing converters.They hold data after the parsing was done whereas it was useless as the data was not accessed outside of the parsing. So this probably consumed extra memory.
This also required we cleaned the data at the beginning of all parsing functions to avoid side effects from previous runs, which made the code harder to read.

As we had global states, we couldn't run json tests in parallel (we currently use `jest --runInBand`) otherwise data clashes would have occurred.
Notice that this had been introduced when using the old json parsing solution that prevented us from instantiating the converters by ourself: global states were the only way we found to share data.
The new custom json parsing solution now allows us to instantiate the converters as we want.

Now we have a processing intermediate data encapsulated in a dedicated class.
No more global state is shared. The instance is created when starting a new parsing and can then be destroyed afterwards.
The encapsulation also reduces the coupling between the converters.