[INFRA] Run GitHub workflows on Ubuntu 20.04
Some workflows were still running on `ubuntu-latest`, which is for now `18.04`
but will be switched to `20.04` soon.
So explicitly use a named version to avoid side effects when the latest version
changes.
In addition and for the same reason, also use the named versions for windows and
macOS currently referenced as the latest version.