[FEAT] Simplify BPMN element identification in DOM (#921)
Assign a dedicated CSS class to BPMN element regarding their `BPMN Shape or Edge kind`. For
instance, all Call Activity nodes will have their class attribute set to `callActivity`.
Possible use cases:
- identify/select BPMN element by using a class selector
- change the BPMN element aspect by cascading styles

In addition, the attribute that let us identify a BPMN element in the DOM has been renamed
from `data-cell-id` to `data-bpmn-id` to have something more convenient with our domain.
As this is not something which is part of our API and we hide in the HTML Element API, there
shouldn't be any impacts.