[INFRA] manage development version and align versions in config files
We now align versions in npm and sonar config files at release time.
So the SonarCloud displays the correct version without us having to maintain
the version manually.

After a new release has been done, we now set the project version with a
dedicated suffix to clearly state that the master branch contains a development
version.
About the implementation
  - it was not possible to use the 'preversion' npm version script otherwise,
  npm would have considered the suffixed version to compute the new one. So, we
  use 2 npm calls when performing the version bump.
  - the suffix removal commit command uses a --no-verify flag to avoid a warning
   generated by husky. This should not cause issue as we don't need linting on
   the files we modified.