[FEAT] Render event-based gateway
Render the exclusive, parallel and default event-based gateway.
The code that generates the cross icons has been refactored to be reused here.
It is now easier to center on the shape and the ratio from parent is more
accurate.
The shapes that already used the cross icon (exclusive and parallel gateways,
cancel event) have been updated accordingly.