[REFACTOR] Rework bpmn-visualization Global Options
The constructor of the lib entry point now takes a single object parameter
including the 'container'.
The container can be passed as an id and the lib will resolve the corresponding
HTML Element.
Property names have been reviewed to better align with the concept described in
the documentation.

As part of this refactoring, the demo.js 'configureBpmnContainer' has been
removed because has not been used anymore since we have introduced tailwind css.