[TEST] Add automatic bundle tests (#1010)
We now have a way to check that bundles are generated and the GitHub workflow
checks this each time the build is run.
This test shows that, in 64aa112d, we have removed by mistake the CommonJS and ESM
bundles generation.

For the IIFE bundle, we also ensure that a test page using the bundle is able to display
a BPMN diagram. The utils code checking the generated SVG is shared with end-to-end tests.
The test is very fast: it doesn't need a dev server to run, the slower part is
the puppeteer startup.
Adding test has also shown that this page hasn't been migrated correctly when the GlobalOptions
have been refactored in 6335a537.