[TEST] Add more BPMN rendering visual tests for labels and waypoints (#1023)
Add new tests to check BPMN rendering for
  - label fonts: ensure the font properties of BPMN Label Styles are used in rendering
  - waypoints: check how the lib behaves when there is no waypoints and when terminal waypoints are defined inside or outside of the source/target shapes

Also always copy test diagrams to destination.
On Ubuntu, not overriding files takes 6ms, overriding 12ms.
The copy is very fast, and not copying the files has side effects: when editing the diagram it is not updated and we don't see the change in tests. This is hard to debug and error prone.