puppeteer waits for the http protocol for the dev server
Puppeteer was previously checking port availability on tcp (default behaviour).
The tcp port is available prior the dev server build finished and it can respond
on http.
So tests started running prior the http server was available, so the 1st ones
got http 404 or unavailable bundle js resource.

tailwindcss: build in sync with the rest of the build. Seems not mandatory here
Previous behaviour: rollup built everything the http server was available, then
the postcss command was run. sometimes, several seconds after the log about the
http server availability. During investigation, we thought it was the cause of
test failures because of resources update.
Consequence of the old behaviour: the latest generated css was not
available or served by the http server because the css copy was done prior the
postcss generation.
The same problem probably occurs with the watch:css (2 client side update
triggered??)