[TEST] Split integration, performance and end-to-end tests (#1056)
The main purpose of this change is to clearly identify the types of the tests.
Previously, a lot of different test types were identified as 'end to end' whereas
they were not.
This led to
  - slower tests: for instance, model and DOM tests don't need Puppeteer, so they were waiting
  for bundle build and http server availability
  - extra filtering (performance along with usual end-to-end tests)
  - extra configuration: visualization tests don't need JSDOM

As part of this change
  - the Jest and Puppeteer configurations have been simplified (remove
redefined defaults, unused mocks)
  - move helper tests code
  - documentation updated accordingly
  - ensure that the SonarCloud coverage computation stays the same


The test split has shown an issue in the end-to-end tests puppeteer configuration
that is now fixed: puppeteer now waits for the http protocol for the dev server.
Puppeteer was previously checking port availability on TCP (default behaviour).
The TCP port is available prior the dev server build finished and it can respond
on http. Then, tests previously started running prior the http server was available,
so the 1st ones got http 404 errors or unavailable bundle js resource.

There was also an issue with tailwindcss: the postcss is now in sync with the rest
of the build.
Previous behaviour: rollup built everything and made the http server was available.
Then, the postcss command was run, sometimes several seconds after the log about the
http server availability.
During investigation about the tests started by jest-puppeteer prior the dev http server
was available, we thought it was the cause of the test failures because of resources update.
Consequence of the old behaviour: the latest generated css was not available or
served by the http server because the css copy was done prior the postcss generation.