[TEST] Generate visual test diff images close to test reports (#1083)
This simplifies test failures review, especially when checking GitHub workflows.
The html test report and the diff images are in the same directory.

The GitHub workflow also upload diff images in the test failures artifact without
extra configuration and there is no more need to dump the diff to the console. So we
also produce smaller and more manageable logs.