[FIX] Don't store 'edge without bpmn element' in the BpmnModel
This is invalid and was not homogeneous with BPMN Shape that are not stored when
we don't get the bpmn element.
In addition, this can produced errors in client code as the bpmn element is not
supposed to be `undefined`.

A dedicated test has been added to manage this use case and parsing tests
that were relying on this 'hole' have been updated to use valid edge
BPMN sources (mainly tests about edge labels).