[REFACTOR] Don't catch and only log errors in the Json Parser
Catching errors and only log them hide parsing problems and generate new ones.
The parsing flow continues after an error is triggered which create side
effects: some valid data are missing and then are not available for the
subsequent parsing process. In that case, the model would be partially filled.
It is so better to propagate the error as it is not managed for now.

In addition, we don't have tests that show we can have such errors so we are not
even sure that such errors could occur.