[REFACTOR] Store part of the BpmnModel into memory for searches (#1115)
This is a first step to avoid depending on the mxGraph model for elements
search.

It introduces a new dedicated registry to manage the BpmnModel and related
objects. They are now used to search element by ids. This first implementation
holds a structure to make search faster: this is probably not optimal and will
be rework later.
The previously called 'DisplayModel' has been extracted out of mxgraph code as
it is not tight to it. It has been renamed into 'RenderedModel' for consistency
with the wording we use everywhere else.

It also simplify some part of the code related to BpmnElement.