[TEST] Add more DOM integration tests (#1193)
We load a file with all supported BPMN elements to ensure we don't have any
issue when producing the SVG.
The code coverage increases significantly as we now pass in all custom `mxGraph`
shapes when running the new tests. For instance, the integration tests coverage
for statements increases from `68.45%` to `93.04%`.