[REFACTOR] Don't use special prefix in private field name (#1196)
The `BpmnVisualization.bpmnModelRegistry` field was previously prefixed by a `_`
as this is sometime done in other projects as a convention to warn about private
fields in JS code. Remember that we don't produce private JS fields at TS
compilation as this not widely supported by browsers, so such fields are
ultimately visible.

This was not consistent with the rest of the code base, so the prefix has been
removed.