[TEST] Improve thresholds configuration for overlay styles visual tests
We previously used a single and high threshold value for all tests. This may
have prevented to detect discrepancies in most of the tests.
Now, we have dedicated thresholds for each tests, so they can be configure to
lowest values and then we will be able to catch more regressions.