[REFACTOR] Improve 'generated svg' e2e tests
Simplify diagram load: the diagram is provided by the dev server. So we now
have a single diagram load method that let remove some utils test code (related
to passing the diagram content as query parameter)
In the 'lib integration test', the diagram content is hard coded in the page.
The page has no access to the regular diagram load method. It is only available
via the demo code and the purpose of the test is to check that the demo code
doesn't interfere with the regular lib code.

Move all code related to the Bpmn Page abstraction in a single file for
consistency.
Introduce a new BpmnPageSvgTester class that permits to align the check done in
the page under test.