[REFACTOR] Make override methods explicit
Use the new TypeScript 4.3 override keyword: `TypeScript will always make sure
that a method with the same name exists in a the base class`.
Make override declaration mandatory: detect method removal in base class,
prevent to accidentally “trample over” a method that exists in a base class.

Also remove some public keywords on methods and constructors for simplicity
(this is the default) and consistency (we mostly avoid public keyword in this
case).