[INFRA] Review the jest configuration (#1467)
We had test path ignore configurations that may filter the whole source code, so Jest failed because no test was found. Also simplify the coverage path filtering.
Path filtering usage is not needed anymore as we use a roots configuration that limit the number of files checks by Jest to find the tests to run.

Past configuration: testPathIgnorePatterns: ['/node_modules/', 'dev', 'dist', 'src'],
This configuration filtered all sources if the project is checkout in a path like ...../development/process-analytics/repo/bpmn-visualization-js.

In addition, check the performance test compilation as part of the build to detect errors early and not only when running the test on local developers machine.