[REFACTOR] Simplify markers computation during rendering
Remove the extra code for the "3 markers case".
We never have 3 markers for now, so the defensive code throwing an error is
never reached.
It will be correctly implemented when managing the initial rendering or the adhoc
subprocess or the compensation marker.

Also simplify the implementation of the 'orderActivityMarkers' function.