[REFACTOR] Simplify markers computation during rendering (#1541)
> information_source For information, in BPMN, it's not possible to have a **loop** marker and a **multi-intance** marker in a same time. But, it's possible to have one of them in combination of the **compensation** marker.
> So the visual tests are wrong.

This was just to demonstrate how the rendering now behaves with 3 markers. I have never had in mind that the provided examples followed the norm. Using a real use case would have required to implement a new marker and I didn't want to loose time on that for such demonstration.