[REFACTOR] Rename ShapeBpmnEventKind into ShapeBpmnEventDefinitionKind
This renaming clarifies the purpose and avoid misunderstandings with the bpmn
element kinds related to BPMN events (start, end, boundary, ...).
The internal model image has been updated accordingly.

For consistency with the type renaming, some changes have been done in the code
to rename `eventKind` into `eventDefinitionKind`.

The renaming may introduce breaking changes for users implementing custom BPMN
Theme, or for people relying on `ShapeUtil` for elements detection.
The impact is very limited. For theming, the touched code is used deeply in the
implementation, so it impacts only people that would have changed the event
icons.
  - StyleUtils: `getBpmnEventKind` method into `getBpmnEventDefinitionKind`
  - ShapeUtil: `topLevelBpmnEventKinds` method into `eventKinds`. We can now use
  a simpler name as we have disambiguated events and event definitions
  - Rename StyleIdentifier BPMN_STYLE_EVENT_KIND into
  BPMN_STYLE_EVENT_DEFINITION_KIND (the value has also been updated for
  consistency)

There are also internal changes that should not impact users
  - internal bpmn model `ShapeBpmnEvent`: rename the class property
  - tests and converters: rename variables, properties, function parameters