[REFACTOR] Simplify parent retrieval in BpmnRenderer
Always returns a parent, there is no need for a dedicated code for boundary
event. If the parent id is set, it is set correctly by the parsing code.
Otherwise, it contains bugs.
We don't have tests that demonstrate the need for such special use case. So
remove this defensive code and use a convenient default.