[REFACTOR] Simplify parent retrieval in BpmnRenderer (#1618)
Always returns a parent, there is no need for a dedicated code for boundary
events. They are supposed to be attached to an existing parent.
More generally, if the parent id is set, it is set correctly by the parsing code.
Otherwise, it means that the parsing code contains bugs.
We don't have tests that demonstrate the need for such special use case. So
remove this defensive code and use a convenient default.