[EXAMPLE] Use deferred script loading
Previously, we put the script declaration at the end of the page body, it
allowed the elimination of parser-blocking JavaScript.
The `defer` attribute is designed for such a purpose.

We only load a single module script in each page, and modules are deferred
automatically. So, we can declare the scripts in the head of the page without
using the defer attribute.

Resources:
  - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Modules#other_differences_between_modules_and_standard_scripts
  - https://developer.mozilla.org/en-US/docs/Web/HTML/Element/script#attr-defer