[FIX] Clean the CSS classes cache on diagram load
Previously, without the cache clean, classes added to a diagram were kept when
loading a new one. So, the DOM element produced after loading a new diagram
could receive an extra class added for the previous diagram.

The cache stored by CssRegistry can now be cleared, and it is cleared on new
diagram load to avoid such an issue.

In addition
  - remove the hack introduced to not face the bug when testing message flow icon
  - add a unit test to demonstrate we cannot accidentally modify the classes
  cache without using the API