[FIX] Correctly apply css classes to message flow icon (#1624)
The 'bpmn-message-flow-icon' class is now always applied. It was previously
applied with a wrong name.
Apply css class when using method of the `BpmnElementsRegistry`. Previously,
the API only applied classes to the message flow related element.

Add integration tests about the SVG nodes in the DOM to check the new behavior.

Update the 'elements-identification.html' example page to demonstrate we can now
style the message flow icon.
It is now possible to not display overlays. This prevents the message flow icon
to be hidden by an overlay.