[FIX] Always apply extra CSS classes to label nodes (#1655)
The extra css classes added via the API weren't applied to label except when
triggering a full diagram redraw i.e. on zoom or panning.

Most of the work done here involved reproducing the issue with integration tests.
They required to tweak the mxGraph configuration as mxGraph was not generating
foreign objects. This seems due to the jsdom environment. Forcing `NO_FO` to
`false` switch the label generation to what we have in the browser.
The integration tests now test the label text and css classes for lane, user
task, sequence flow, start and end events.

In MxCellUpdater, remove call to `state.shape.apply(state)`
This code applied the style of the given `mxCellState` to the shape and only
default styles known by mxGraph.
The 'extra css style' is not taken from the shape but from the state. See the
overridden implementation of `mxgraph.mxShape.prototype.createSvgCanvas` defined
in `ShapeConfigurator` for more details.
All tests pass after the removal, without changes in tests.
Manual tests performed on zoom and panning with the elements-identification.html
test pages are still ok (the CSS classes are still in the DOM).

Additional changes:
  - Add information about the svg groups created by mxgraph
  - Update the `query selectors` used in tests. Simplify definition and maximize
  share between integration and e2e tests to simplify maintenance.
  - html-utils.ts: refactor to avoid duplications. More refactor will be done
  in the future.
  - Add css style for labels in elements-identification.html and also detect Pool