[FEAT] Generate more precise CSS classes on BPMN elements
This let style elements of the same type, for instance activity, container (lane
or pool), event, fow, gateway or task.
A specific class is also generated for message flow icon regarding it is related
to an `initiating` or `non-initiating` message flow.

The `elements-identifications.html` example now uses the 'type' css classes for
simplification and the new icon css class to demonstrate its usage.