[TEST] Split mxGraph model tests
The existing test file was too large, so it was hard to maintain and navigate.
The BPMN elements checks were done in a single test. It wasn't possible to run a
part of the checks and if one check failed, we couldn't know if other elements
had errors as the test stopped at the first encountered error.

Included changes
  - Move overlays tests to a dedicated file.
  - Introduce a test for each group BPMN types or kinds. Instead of single test,
   we now have 38 tests.

Also test the robustness of the lib: test BpmnVisualization constructor when the
`options` parameter is `null`.