[INFRA] Investigate flaky bundle tests (#1700)
We now set launch timeout to 60 sec on macOS when running on CI. Also activate the playwright browser logs
to measure the time taken for browser launch. Even if this doesn't fix the problem, we will have more insights.

The playwright default launch timeout is 30 seconds. On macOS, we sometimes see timeout errors when running
GitHub Actions, so increase the timeout.