[INFRA] Run tests also with Chrome and Edge (#1701)
Run bundle and e2e tests with Chrome on all OS and with Edge on Windows only.
Now only run bundle tests with WebKit on macOS.
Reduce the duplication in the jest-playwright configuration by introducing shared functions.

Also improve the browser launch timeout configuration to avoid false positive when running on GitHub Actions.
On CI, always increase this timeout when running on macOS and Windows: the hosts using these OS
are slower than the ones on Ubuntu. So, quite often, we have test failures because of browser launch timeout
with the default value.
For other environment (in particular, local machine) or GH Actions running Ubuntu, timeout doesn't
occur, so we keep the default: this will let us detect any slowness that could be introduced in the future.