[DOC] Reference the BPMN Support doc instead of source in README
The README is bundled in the npm package. In that case, the other documentation
source files are not available. There are online services that use the README
as the home page of the library (npmjs, skypack and others). In that case, the
link to the BPMN Support documentation is broken.
The direct use of the documentation url fixes the issue.