[FEAT] Improve the robustness of the XML Parser
Generate a clearer error message when the underlying `fast-xml-parser` 'parse'
function is unable to parse the BPMN source.
For instance, passing something like png, pdf or zip content make the 'parse'
function throw an Error as it tries to access to property of undefined instance.
We now catch the error and return a convenient error message.