[FIX] Do not display window alert on BPMN load error (#1761)
Alert is something we may want to have in the example/demo pages of the library,
but not in the library itself.
It is the responsibility of the integration to manage the error as it wants.
Previously, alert popups were showing all the time and no one wanted that in
their application.

The window alert is now used only in the demo/example code, so the behavior
remains unchanged in the related pages. Also open alert for other kind of errors
(diagram fetch for instance) in the same pages for consistency.
Error management could be improved later in these pages.

Nice side effects in integration tests: there is less error in the test logs.
The window.alert function doesn't exist in the JSDOM environment so calls to the
function previously generate disturbing log errors.