[TEST] Add more visual tests for terminal waypoints special cases
Add special waypoint positions test cases
  - none/outside for associations (inside already existed)
  - none/inside/outside for message flows

Rename existing BPMN diagram files for consistency.
For some existing diagrams, remove the exporter information (not relevant here).

Fix the diagram about 'message flows with complex path'. A source terminal point
was outside the shape and this is not something we want to test with this
diagram (we have a special 'waypoints outside shapes' diagram for that).