[TEST] Improve the e2e test reports (#1785)
Implement a jest-image-snapshot custom matcher to provide more information on error. It makes the debug easier.
  - Generate expected and actual images along with the already existing diff image. We can then make the diff with these images with other tools
  - Display the configured 'failureThreshold' in the error message for comparison with the actual diff percentage.

It could be integrated directly in jest-image-snapshot, but currently, the project seems inactive and doesn't merge PR. So do it here for now. We will propose the change later.

Introduce usage of `jest-html-reporters` to let display diff, expected and actual images directly in an HTML page.
The previously used HTML report is kept to quickly access to the diff percentage: it make updating the thresholds
easy and we want to keep that.