[FIX] Fix the type definitions (#1814)
In the npm package, the definitions were previously located in dist/src instead of in dist/. It also contained development/demo types in dist/dev. So, project integrating bpmn-visualization were unable to retrieve the types and got errors like TS7016: Could not find a declaration file for module 'bpmn-visualization'

The issue was located in the rollup configuration that override the tsconfig file when building the npm package. The 'include' configuration was not correctly overridden, so we used the development values.
This is now fixed and there is a new test that will prevent this from happening again.

Some definition files previously referenced types or classes that were not defined. Such elements are marked as internal and strip from the definitions.
So this was generating errors like TS1192 or TS2305 (mainly involving BpmnModel, Edge or Shape).
There were some functions and classes that weren't marked as `@internal` and so created the broken references.

To improve the usage of bpmn-visualization in TypeScript projects, the README now provides more information about the TypeScript support and how to use the lib in TS projects.