[EXAMPLE] Display the version in the footer
Use the new Version API to display the bpmn-visualization and mxGraph versions
in the footer.
The general layout of the page has been updated to simplify the addition of the
footer using flex containers.
It also includes initial responsive improvements: when using small screens, the
bpmn-container has no rounded border, block are stacked vertically without
horizontal overflow).

The bpmn-container now grows in the vertical axis and fill the whole white panel
(not small s screens).
The dimensions are also fully managed by CSS, there is no more javascript
function doing some magic.

The drop-container feature has been updated to avoid hiding the bpmn-container
in the demo page. The text is now fully center both horizontally and vertically
whatever the height and width are.
It also works in the other pages that are not responsive.

Refactor
  - rename demo.js into index.js for consistency with the page name
  - remove the `configureControlPanel` js function as it was not used anymore.
  It was a relic of the 1st demo page prior we introduce a new design and create
  specific test pages that have controls.
  - index.html: remove some unused CSS classes