[INFRA] Do not fail when surge teardown fails
When a Pull Request is created by dependabot or an external contributor, the
deployments are not created. But when the PR is merged or closed by a repository
collaborator, the teardown is triggered. As there is no deployment, the teardown
fails.
Previously, this mark the workflow execution in error. Now, we don't fail on
error anymore because we don't really care.
We have a scheduled task that regularly teardown old surge deployments, so we
know they will be removed later.